refactor(1015): extract DeviceFetchConfig to src/refactors/device_data_fetcher.py (T-01, Cat E)
